HeroBoard.c
This is a Pebble app keyboard based on the game Guitar Hero. 
It is NOT a final product/watchapp. It is meant to be used with T3/T9 text prediction scripts. 
When completed, it could be implemented as an exclusive keyboard for Pebble. 
It was developed by rodrigo@mastria.com.br as part of the HackIllinois.org 2015